refactor: extract settings tabs into registry pattern
