+++ changelog/unreleased/ocp-settings-tab-registry.md
@@ -1,0 +1,2 @@
+### Changed
+- **Refactored settings dialog** — Extracted 6 settings tabs into separate files with a registry pattern for easier extensibility
