+++ src/components/SettingsDialog.ts
@@ -1,39 +1,6 @@
-import { invoke } from '@tauri-apps/api/core';
-import { revealItemInDir } from '@tauri-apps/plugin-opener';
-import {
-  keybindingStore,
-  DEFAULT_SHORTCUTS,
-  formatChord,
-  eventToChord,
-  type ActionId,
-  type ShortcutCategory,
-} from '../state/keybinding-store';
-import { notificationStore } from '../state/notification-store';
-import { terminalSettingsStore } from '../state/terminal-settings-store';
 import { settingsTabStore } from '../state/settings-tab-store';
-import { workspaceService } from '../services/workspace-service';
-import { playNotificationSound, type SoundPreset } from '../services/notification-sound';
-import { remoteSettingsStore, generatePassword, generateApiKey } from '../state/remote-settings-store';
-import { themeStore } from '../state/theme-store';
-import type { ThemeDefinition } from '../themes/types';
-import { createThemePreview } from './ThemePreview';
-import type { ShellType } from '../state/store';
-import { showFileEditorDialog } from './FileEditorDialog';
-import { getPluginRegistry } from '../plugins/index';
-import { pluginStore } from '../plugins/plugin-store';
-import { pluginInstaller } from '../plugins/plugin-installer';
-import { createPluginCard } from './PluginCard';
-import type { RegistryEntry, PluginRegistryData } from '../plugins/types';
-import registryData from '../plugins/registry.json';
-
-function formatCustomSoundName(filename: string): string {
-  // Strip extension
-  const name = filename.replace(/\.[^.]+$/, '');
-  // Replace _ and - with spaces, then title-case
-  return name
-    .replace(/[_-]/g, ' ')
-    .replace(/\b\w/g, c => c.toUpperCase());
-}
+import { getSettingsTabRegistry } from './settings/index';
+import type { SettingsDialogContext } from './settings/types';
 
 /**
  * Show the settings dialog for customising themes, notifications, and
@@ -59,15 +26,12 @@ export function showSettingsDialog(): Promise<void> {
 
     dialog.appendChild(header);
 
-    // ── Tab bar ──────────────────────────────────────────────────
-    const allTabs: Record<string, string> = {
-      themes: 'Themes',
-      terminal: 'Terminal',
-      notifications: 'Notifications',
-      plugins: 'Plugins',
-      shortcuts: 'Shortcuts',
-      remote: 'Remote',
-    };
+    // ── Tab registry ────────────────────────────────────────────
+    const registry = getSettingsTabRegistry();
+    const allTabs: Record<string, string> = {};
+    for (const provider of registry) {
+      allTabs[provider.id] = provider.label;
+    }
 
     let tabOrder = settingsTabStore.getTabOrder();
     let activeTab = tabOrder[0];
@@ -78,9 +42,20 @@ export function showSettingsDialog(): Promise<void> {
     const tabButtons: Record<string, HTMLButtonElement> = {};
     const tabContents: Record<string, HTMLDivElement> = {};
 
+    const dialogContext: SettingsDialogContext = {
+      renderTabBar: () => buildTabBar(),
+    };
+
+    // Build tab contents from registry
+    for (const provider of registry) {
+      const content = provider.buildContent(dialogContext);
+      tabContents[provider.id] = content;
+    }
+
     function buildTabBar() {
       tabBar.textContent = '';
       for (const id of tabOrder) {
+        if (!allTabs[id]) continue;
         let btn = tabButtons[id];
         if (!btn) {
           btn = document.createElement('button');
@@ -118,7 +93,6 @@ export function showSettingsDialog(): Promise<void> {
           }
           if (!dragging) return;
 
-          // Determine which tab the pointer is over
           for (const id of tabOrder) {
             const btn = tabButtons[id];
             if (!btn || id === dragTabId) continue;
@@ -160,1257 +134,14 @@ export function showSettingsDialog(): Promise<void> {
       }
     }
 
-    // ── Themes tab content ──────────────────────────────────────
-    const themesContent = document.createElement('div');
-    themesContent.className = 'settings-tab-content active';
-    tabContents['themes'] = themesContent;
-
-    const themeGrid = document.createElement('div');
-    themeGrid.className = 'theme-grid';
-    themesContent.appendChild(themeGrid);
-
-    function renderThemeGrid() {
-      themeGrid.textContent = '';
-      const allThemes = themeStore.getAllThemes();
-      const activeTheme = themeStore.getActiveTheme();
-
-      for (const theme of allThemes) {
-        const card = document.createElement('div');
-        card.className = 'theme-card' + (theme.id === activeTheme.id ? ' active' : '');
-
-        // Canvas preview
-        const preview = createThemePreview(theme, 280, 140);
-        card.appendChild(preview);
-
-        // Info area
-        const info = document.createElement('div');
-        info.className = 'theme-card-info';
-
-        const nameEl = document.createElement('div');
-        nameEl.className = 'theme-card-name';
-        nameEl.textContent = theme.name;
-        info.appendChild(nameEl);
-
-        const descEl = document.createElement('div');
-        descEl.className = 'theme-card-description';
-        descEl.textContent = theme.description;
-        info.appendChild(descEl);
-
-        const authorEl = document.createElement('div');
-        authorEl.className = 'theme-card-author';
-        authorEl.textContent = theme.author;
-        info.appendChild(authorEl);
-
-        card.appendChild(info);
-
-        // Remove button for non-builtin themes
-        if (!theme.builtin) {
-          const actions = document.createElement('div');
-          actions.className = 'theme-card-actions';
-
-          const removeBtn = document.createElement('button');
-          removeBtn.className = 'dialog-btn dialog-btn-secondary';
-          removeBtn.textContent = 'Remove';
-          removeBtn.style.fontSize = '11px';
-          removeBtn.style.padding = '2px 10px';
-          removeBtn.onclick = (e) => {
-            e.stopPropagation();
-            themeStore.removeCustomTheme(theme.id);
-            renderThemeGrid();
-          };
-          actions.appendChild(removeBtn);
-          card.appendChild(actions);
-        }
-
-        card.onclick = () => {
-          themeStore.setActiveTheme(theme.id);
-          renderThemeGrid();
-        };
-
-        themeGrid.appendChild(card);
+    // Append tab contents to dialog (first tab is active)
+    for (const provider of registry) {
+      const content = tabContents[provider.id];
+      if (provider.id === activeTab) {
+        content.classList.add('active');
       }
+      dialog.appendChild(content);
     }
-
-    renderThemeGrid();
-
-    // Import button
-    const importBtn = document.createElement('button');
-    importBtn.className = 'dialog-btn dialog-btn-secondary theme-import-btn';
-    importBtn.textContent = 'Import Theme (JSON)';
-    importBtn.onclick = () => {
-      const fileInput = document.createElement('input');
-      fileInput.type = 'file';
-      fileInput.accept = '.json';
-      fileInput.style.display = 'none';
-      fileInput.onchange = async () => {
-        const file = fileInput.files?.[0];
-        if (!file) return;
-        try {
-          const text = await file.text();
-          const parsed = JSON.parse(text) as ThemeDefinition;
-          // Validate required fields
-          if (
-            !parsed.id ||
-            !parsed.name ||
-            !parsed.terminal ||
-            !parsed.ui
-          ) {
-            alert('Invalid theme file: missing required fields (id, name, terminal, ui).');
-            return;
-          }
-          parsed.builtin = false;
-          themeStore.addCustomTheme(parsed);
-          renderThemeGrid();
-        } catch (err) {
-          alert('Failed to import theme: ' + (err instanceof Error ? err.message : String(err)));
-        }
-        fileInput.remove();
-      };
-      document.body.appendChild(fileInput);
-      fileInput.click();
-    };
-    themesContent.appendChild(importBtn);
-
-    dialog.appendChild(themesContent);
-
-    // ── Terminal tab content ─────────────────────────────────────
-    const terminalContent = document.createElement('div');
-    terminalContent.className = 'settings-tab-content';
-    tabContents['terminal'] = terminalContent;
-
-    const termSection = document.createElement('div');
-    termSection.className = 'settings-section';
-
-    const termTitle = document.createElement('div');
-    termTitle.className = 'settings-section-title';
-    termTitle.textContent = 'Default Shell';
-    termSection.appendChild(termTitle);
-
-    const termDesc = document.createElement('div');
-    termDesc.className = 'settings-description';
-    termDesc.textContent = 'Choose which shell to use when creating new terminals and workspaces.';
-    termSection.appendChild(termDesc);
-
-    interface ShellOption {
-      id: string;
-      label: string;
-      tooltip: string;
-      shellType: ShellType;
-      checkAvailability?: () => Promise<boolean>;
-    }
-
-    const shellOptions: ShellOption[] = [
-      {
-        id: 'windows',
-        label: 'PowerShell',
-        tooltip: 'Built-in on all Windows. Rich scripting, but slower startup (~1-2s).',
-        shellType: { type: 'windows' },
-      },
-      {
-        id: 'pwsh',
-        label: 'PowerShell 7',
-        tooltip: 'Modern, faster startup, cross-platform. Must be installed separately.',
-        shellType: { type: 'pwsh' },
-        checkAvailability: () => invoke<boolean>('is_pwsh_available'),
-      },
-      {
-        id: 'cmd',
-        label: 'Command Prompt',
-        tooltip: 'Fastest startup. Lightweight but limited scripting capabilities.',
-        shellType: { type: 'cmd' },
-      },
-      {
-        id: 'wsl',
-        label: 'WSL',
-        tooltip: 'Full Linux environment. Requires WSL to be installed.',
-        shellType: { type: 'wsl' },
-        checkAvailability: () => workspaceService.isWslAvailable(),
-      },
-      {
-        id: 'custom',
-        label: 'Custom Shell',
-        tooltip: 'Run any shell executable (e.g. nu.exe, fish, bash.exe).',
-        shellType: { type: 'custom', program: '' },
-      },
-    ];
-
-    const currentDefault = terminalSettingsStore.getDefaultShell();
-    let selectedShellId: string = currentDefault.type;
-
-    const radioGroup = document.createElement('div');
-    radioGroup.className = 'shell-radio-group';
-
-    // WSL distribution dropdown (hidden unless WSL is selected)
-    const wslDistroRow = document.createElement('div');
-    wslDistroRow.className = 'shell-wsl-distro-row';
-    wslDistroRow.style.display = 'none';
-
-    // Custom shell inputs (hidden unless Custom is selected)
-    const customShellRow = document.createElement('div');
-    customShellRow.className = 'shell-wsl-distro-row';
-    customShellRow.style.display = 'none';
-
-    const customProgramLabel = document.createElement('span');
-    customProgramLabel.className = 'shortcut-label';
-    customProgramLabel.textContent = 'Program';
-    customShellRow.appendChild(customProgramLabel);
-
-    const customProgramInput = document.createElement('input');
-    customProgramInput.type = 'text';
-    customProgramInput.className = 'notification-preset';
-    customProgramInput.placeholder = 'e.g. nu.exe, fish, bash.exe';
-    customProgramInput.style.flex = '1';
-    customShellRow.appendChild(customProgramInput);
-
-    const customArgsLabel = document.createElement('span');
-    customArgsLabel.className = 'shortcut-label';
-    customArgsLabel.textContent = 'Args';
-    customArgsLabel.style.marginLeft = '8px';
-    customShellRow.appendChild(customArgsLabel);
-
-    const customArgsInput = document.createElement('input');
-    customArgsInput.type = 'text';
-    customArgsInput.className = 'notification-preset';
-    customArgsInput.placeholder = 'e.g. -l --config';
-    customArgsInput.style.flex = '1';
-    customShellRow.appendChild(customArgsInput);
-
-    // Pre-populate if current default is custom
-    if (currentDefault.type === 'custom') {
-      const custom = currentDefault as { type: 'custom'; program: string; args?: string[] };
-      customProgramInput.value = custom.program;
-      customArgsInput.value = (custom.args ?? []).join(' ');
-      customShellRow.style.display = 'flex';
-    }
-
-    // Custom shell input change handlers
-    const updateCustomShell = () => {
-      if (selectedShellId === 'custom' && customProgramInput.value.trim()) {
-        const args = customArgsInput.value.trim()
-          ? customArgsInput.value.trim().split(/\s+/)
-          : undefined;
-        terminalSettingsStore.setDefaultShell({
-          type: 'custom',
-          program: customProgramInput.value.trim(),
-          args,
-        });
-      }
-    };
-    customProgramInput.onchange = updateCustomShell;
-    customArgsInput.onchange = updateCustomShell;
-
-    const wslDistroLabel = document.createElement('span');
-    wslDistroLabel.className = 'shortcut-label';
-    wslDistroLabel.textContent = 'Distribution';
-    wslDistroRow.appendChild(wslDistroLabel);
-
-    const wslDistroSelect = document.createElement('select');
-    wslDistroSelect.className = 'notification-preset';
-    wslDistroRow.appendChild(wslDistroSelect);
-
-    // Track availability per option
-    const optionAvailability: Record<string, boolean> = {};
-    const optionElements: Record<string, HTMLDivElement> = {};
-
-    for (const opt of shellOptions) {
-      const row = document.createElement('div');
-      row.className = 'shell-option-row';
-
-      const radio = document.createElement('input');
-      radio.type = 'radio';
-      radio.name = 'default-shell';
-      radio.id = `shell-${opt.id}`;
-      radio.value = opt.id;
-      radio.checked = selectedShellId === opt.id;
-      row.appendChild(radio);
-
-      const label = document.createElement('label');
-      label.htmlFor = `shell-${opt.id}`;
-      label.className = 'shell-option-label';
-
-      const nameSpan = document.createElement('span');
-      nameSpan.className = 'shell-option-name';
-      nameSpan.textContent = opt.label;
-      label.appendChild(nameSpan);
-
-      // Info tooltip icon
-      const infoIcon = document.createElement('span');
-      infoIcon.className = 'shell-info-icon';
-      infoIcon.textContent = '\u24D8'; // circled i
-      infoIcon.title = opt.tooltip;
-      label.appendChild(infoIcon);
-
-      // "(not installed)" label placeholder
-      const unavailableLabel = document.createElement('span');
-      unavailableLabel.className = 'shell-unavailable-label';
-      unavailableLabel.style.display = 'none';
-      unavailableLabel.textContent = '(not installed)';
-      label.appendChild(unavailableLabel);
-
-      row.appendChild(label);
-
-      radio.onchange = () => {
-        if (!radio.checked) return;
-        selectedShellId = opt.id;
-
-        // Show/hide conditional rows
-        wslDistroRow.style.display = opt.id === 'wsl' ? 'flex' : 'none';
-        customShellRow.style.display = opt.id === 'custom' ? 'flex' : 'none';
-
-        if (opt.id === 'wsl') {
-          const distro = wslDistroSelect.value || undefined;
-          terminalSettingsStore.setDefaultShell({ type: 'wsl', distribution: distro });
-        } else if (opt.id === 'custom') {
-          if (customProgramInput.value.trim()) {
-            const args = customArgsInput.value.trim()
-              ? customArgsInput.value.trim().split(/\s+/)
-              : undefined;
-            terminalSettingsStore.setDefaultShell({
-              type: 'custom',
-              program: customProgramInput.value.trim(),
-              args,
-            });
-          }
-        } else {
-          terminalSettingsStore.setDefaultShell(opt.shellType);
-        }
-      };
-
-      radioGroup.appendChild(row);
-      optionElements[opt.id] = row;
-    }
-
-    termSection.appendChild(radioGroup);
-    termSection.appendChild(wslDistroRow);
-    termSection.appendChild(customShellRow);
-
-    // WSL distro select change handler
-    wslDistroSelect.onchange = () => {
-      if (selectedShellId === 'wsl') {
-        const distro = wslDistroSelect.value || undefined;
-        terminalSettingsStore.setDefaultShell({ type: 'wsl', distribution: distro });
-      }
-    };
-
-    // Check availability asynchronously and disable unavailable options
-    for (const opt of shellOptions) {
-      if (!opt.checkAvailability) {
-        optionAvailability[opt.id] = true;
-        continue;
-      }
-      opt.checkAvailability().then(available => {
-        optionAvailability[opt.id] = available;
-        if (!available) {
-          const row = optionElements[opt.id];
-          const radio = row.querySelector('input') as HTMLInputElement;
-          radio.disabled = true;
-          row.classList.add('disabled');
-          const unavailableSpan = row.querySelector('.shell-unavailable-label') as HTMLElement;
-          if (unavailableSpan) unavailableSpan.style.display = 'inline';
-        }
-      }).catch(() => {
-        // Treat check failure as available
-        optionAvailability[opt.id] = true;
-      });
-    }
-
-    // If WSL is currently selected, show the distro dropdown and load distributions
-    if (currentDefault.type === 'wsl') {
-      wslDistroRow.style.display = 'flex';
-      workspaceService.getWslDistributions().then(distros => {
-        wslDistroSelect.textContent = '';
-        const defaultOpt = document.createElement('option');
-        defaultOpt.value = '';
-        defaultOpt.textContent = 'Default';
-        wslDistroSelect.appendChild(defaultOpt);
-
-        for (const d of distros) {
-          const dOpt = document.createElement('option');
-          dOpt.value = d;
-          dOpt.textContent = d;
-          if ((currentDefault as { type: 'wsl'; distribution?: string }).distribution === d) {
-            dOpt.selected = true;
-          }
-          wslDistroSelect.appendChild(dOpt);
-        }
-      }).catch(() => {});
-    } else {
-      // Pre-load distros so they're ready if user switches to WSL
-      workspaceService.getWslDistributions().then(distros => {
-        wslDistroSelect.textContent = '';
-        const defaultOpt = document.createElement('option');
-        defaultOpt.value = '';
-        defaultOpt.textContent = 'Default';
-        wslDistroSelect.appendChild(defaultOpt);
-
-        for (const d of distros) {
-          const dOpt = document.createElement('option');
-          dOpt.value = d;
-          dOpt.textContent = d;
-          wslDistroSelect.appendChild(dOpt);
-        }
-      }).catch(() => {});
-    }
-
-    terminalContent.appendChild(termSection);
-
-    // ── Scrollback section ─────────────────────────────────────
-    const scrollSection = document.createElement('div');
-    scrollSection.className = 'settings-section';
-
-    const scrollTitle = document.createElement('div');
-    scrollTitle.className = 'settings-section-title';
-    scrollTitle.textContent = 'Scrollback';
-    scrollSection.appendChild(scrollTitle);
-
-    const autoScrollRow = document.createElement('div');
-    autoScrollRow.className = 'shortcut-row';
-    const autoScrollLabel = document.createElement('span');
-    autoScrollLabel.className = 'shortcut-label';
-    autoScrollLabel.textContent = 'Auto-scroll to bottom on new output';
-    autoScrollRow.appendChild(autoScrollLabel);
-    const autoScrollCheckbox = document.createElement('input');
-    autoScrollCheckbox.type = 'checkbox';
-    autoScrollCheckbox.className = 'notification-checkbox';
-    autoScrollCheckbox.checked = terminalSettingsStore.getAutoScrollOnOutput();
-    autoScrollCheckbox.onchange = () => {
-      terminalSettingsStore.setAutoScrollOnOutput(autoScrollCheckbox.checked);
-    };
-    autoScrollRow.appendChild(autoScrollCheckbox);
-    scrollSection.appendChild(autoScrollRow);
-
-    const autoScrollDesc = document.createElement('div');
-    autoScrollDesc.className = 'settings-description';
-    autoScrollDesc.textContent = 'When enabled, new terminal output will snap the view to the bottom even while you are scrolled up. When disabled (default), your scroll position is preserved.';
-    scrollSection.appendChild(autoScrollDesc);
-
-    terminalContent.appendChild(scrollSection);
-
-    // ── CMD Aliases section ────────────────────────────────────
-    const aliasSection = document.createElement('div');
-    aliasSection.className = 'settings-section';
-
-    const aliasTitle = document.createElement('div');
-    aliasTitle.className = 'settings-section-title';
-    aliasTitle.textContent = 'CMD Aliases';
-    aliasSection.appendChild(aliasTitle);
-
-    const aliasDesc = document.createElement('div');
-    aliasDesc.className = 'settings-description';
-    aliasDesc.textContent = 'Define command aliases for Command Prompt sessions. Aliases load automatically via the Windows AutoRun registry key.';
-    aliasSection.appendChild(aliasDesc);
-
-    const aliasRow = document.createElement('div');
-    aliasRow.className = 'shortcut-row';
-
-    const aliasLabel = document.createElement('span');
-    aliasLabel.className = 'shortcut-label';
-    aliasLabel.textContent = 'Aliases file';
-    aliasRow.appendChild(aliasLabel);
-
-    const aliasStatus = document.createElement('span');
-    aliasStatus.className = 'shell-info-icon';
-    aliasStatus.style.marginLeft = '8px';
-    aliasStatus.style.fontSize = '11px';
-    aliasStatus.style.opacity = '0.7';
-
-    const editAliasBtn = document.createElement('button');
-    editAliasBtn.className = 'dialog-btn dialog-btn-secondary';
-    editAliasBtn.textContent = 'Edit Aliases';
-    editAliasBtn.onclick = async () => {
-      try {
-        const aliasPath = await invoke<string>('get_cmd_aliases_path');
-        const defaultTemplate = '@echo off\ndoskey dclaude=claude --dangerously-skip-permissions $*\n';
-        await showFileEditorDialog('CMD Aliases', aliasPath, defaultTemplate);
-        // After editor closes, ensure registry is configured
-        const status = await invoke<string>('ensure_cmd_autorun');
-        if (status === 'configured' || status === 'appended') {
-          aliasStatus.textContent = 'AutoRun configured';
-        } else {
-          aliasStatus.textContent = 'Already configured';
-        }
-      } catch (err) {
-        console.warn('CMD aliases setup failed:', err);
-        aliasStatus.textContent = 'Setup failed';
-      }
-    };
-    aliasRow.appendChild(editAliasBtn);
-    aliasRow.appendChild(aliasStatus);
-    aliasSection.appendChild(aliasRow);
-
-    terminalContent.appendChild(aliasSection);
-
-    dialog.appendChild(terminalContent);
-
-    // ── Notifications tab content ───────────────────────────────
-    const notifContent = document.createElement('div');
-    notifContent.className = 'settings-tab-content';
-    tabContents['notifications'] = notifContent;
-
-    const notifSection = document.createElement('div');
-    notifSection.className = 'settings-section';
-
-    const notifTitle = document.createElement('div');
-    notifTitle.className = 'settings-section-title';
-    notifTitle.textContent = 'Notifications';
-    notifSection.appendChild(notifTitle);
-
-    // Enable/disable row
-    const enableRow = document.createElement('div');
-    enableRow.className = 'shortcut-row';
-    const enableLabel = document.createElement('span');
-    enableLabel.className = 'shortcut-label';
-    enableLabel.textContent = 'Sound notifications';
-    enableRow.appendChild(enableLabel);
-    const enableCheckbox = document.createElement('input');
-    enableCheckbox.type = 'checkbox';
-    enableCheckbox.className = 'notification-checkbox';
-    enableCheckbox.checked = notificationStore.getSettings().globalEnabled;
-    enableCheckbox.onchange = () => {
-      notificationStore.setGlobalEnabled(enableCheckbox.checked);
-    };
-    enableRow.appendChild(enableCheckbox);
-    notifSection.appendChild(enableRow);
-
-    // Volume row
-    const volumeRow = document.createElement('div');
-    volumeRow.className = 'shortcut-row';
-    const volumeLabel = document.createElement('span');
-    volumeLabel.className = 'shortcut-label';
-    volumeLabel.textContent = 'Volume';
-    volumeRow.appendChild(volumeLabel);
-    const volumeSlider = document.createElement('input');
-    volumeSlider.type = 'range';
-    volumeSlider.className = 'notification-volume';
-    volumeSlider.min = '0';
-    volumeSlider.max = '100';
-    volumeSlider.value = String(Math.round(notificationStore.getSettings().volume * 100));
-    volumeSlider.oninput = () => {
-      notificationStore.setVolume(parseInt(volumeSlider.value) / 100);
-    };
-    volumeRow.appendChild(volumeSlider);
-    notifSection.appendChild(volumeRow);
-
-    // Sound preset row
-    const presetRow = document.createElement('div');
-    presetRow.className = 'shortcut-row';
-    const presetLabel = document.createElement('span');
-    presetLabel.className = 'shortcut-label';
-    presetLabel.textContent = 'Sound';
-    presetRow.appendChild(presetLabel);
-    const presetSelect = document.createElement('select');
-    presetSelect.className = 'notification-preset';
-
-    // Built-in presets
-    const builtinGroup = document.createElement('optgroup');
-    builtinGroup.label = 'Built-in Sounds';
-    const presets: { value: SoundPreset; label: string }[] = [
-      { value: 'chime', label: 'Chime' },
-      { value: 'bell', label: 'Bell' },
-      { value: 'ping', label: 'Ping' },
-      { value: 'soft-rise', label: 'Soft Rise' },
-      { value: 'crystal', label: 'Crystal' },
-      { value: 'bubble', label: 'Bubble' },
-      { value: 'harp', label: 'Harp' },
-      { value: 'marimba', label: 'Marimba' },
-      { value: 'cosmic', label: 'Cosmic' },
-      { value: 'droplet', label: 'Droplet' },
-    ];
-    presets.forEach(p => {
-      const opt = document.createElement('option');
-      opt.value = p.value;
-      opt.textContent = p.label;
-      if (p.value === notificationStore.getSettings().soundPreset) opt.selected = true;
-      builtinGroup.appendChild(opt);
-    });
-    presetSelect.appendChild(builtinGroup);
-
-    // Load custom sounds and populate dropdown
-    const customGroup = document.createElement('optgroup');
-    customGroup.label = 'Custom Sounds';
-    invoke<string[]>('list_custom_sounds').then(files => {
-      if (files.length === 0) return;
-      const currentPreset = notificationStore.getSettings().soundPreset;
-      files.forEach(filename => {
-        const opt = document.createElement('option');
-        opt.value = `custom:${filename}`;
-        opt.textContent = formatCustomSoundName(filename);
-        if (`custom:${filename}` === currentPreset) opt.selected = true;
-        customGroup.appendChild(opt);
-      });
-      presetSelect.appendChild(customGroup);
-    }).catch(() => {
-      // Custom sounds unavailable — silently omit the group
-    });
-
-    presetSelect.onchange = () => {
-      const selected = presetSelect.value as SoundPreset;
-      notificationStore.setSoundPreset(selected);
-      playNotificationSound(selected, notificationStore.getSettings().volume);
-    };
-    presetRow.appendChild(presetSelect);
-
-    const testBtn = document.createElement('button');
-    testBtn.className = 'dialog-btn dialog-btn-secondary';
-    testBtn.textContent = 'Test';
-    testBtn.style.marginLeft = '8px';
-    testBtn.onclick = () => {
-      const s = notificationStore.getSettings();
-      playNotificationSound(s.soundPreset, s.volume);
-    };
-    presetRow.appendChild(testBtn);
-    notifSection.appendChild(presetRow);
-
-    // Custom sounds folder row
-    const folderRow = document.createElement('div');
-    folderRow.className = 'shortcut-row';
-    const folderLabel = document.createElement('span');
-    folderLabel.className = 'shortcut-label';
-    folderLabel.textContent = 'Custom sounds';
-    const customSoundsInfo = document.createElement('span');
-    customSoundsInfo.className = 'shell-info-icon';
-    customSoundsInfo.textContent = '\u24D8';
-    customSoundsInfo.title =
-      'Drop .wav, .mp3, or .ogg files into this folder to add them as notification sound options.';
-    folderLabel.appendChild(customSoundsInfo);
-    folderRow.appendChild(folderLabel);
-    const openFolderBtn = document.createElement('button');
-    openFolderBtn.className = 'dialog-btn dialog-btn-secondary';
-    openFolderBtn.textContent = 'Open Sounds Folder';
-    openFolderBtn.onclick = async () => {
-      try {
-        const dir: string = await invoke('get_sounds_dir');
-        await revealItemInDir(dir);
-      } catch (e) {
-        console.warn('Failed to open sounds folder:', e);
-      }
-    };
-    folderRow.appendChild(openFolderBtn);
-    notifSection.appendChild(folderRow);
-
-    // Idle activity notifications toggle
-    const idleRow = document.createElement('div');
-    idleRow.className = 'shortcut-row';
-    const idleLabel = document.createElement('span');
-    idleLabel.className = 'shortcut-label';
-    idleLabel.textContent = 'Idle activity notifications';
-    const idleInfo = document.createElement('span');
-    idleInfo.className = 'shell-info-icon';
-    idleInfo.textContent = '\u24D8';
-    idleInfo.title =
-      'Plays a sound when a background terminal stops producing output (e.g. a long build finishes or a CLI tool is waiting for input). Only triggers for terminals you are not currently looking at.';
-    idleLabel.appendChild(idleInfo);
-    idleRow.appendChild(idleLabel);
-    const idleCheckbox = document.createElement('input');
-    idleCheckbox.type = 'checkbox';
-    idleCheckbox.className = 'notification-checkbox';
-    idleCheckbox.checked = notificationStore.getSettings().idleNotifyEnabled;
-    idleCheckbox.onchange = () => {
-      notificationStore.setIdleNotifyEnabled(idleCheckbox.checked);
-    };
-    idleRow.appendChild(idleCheckbox);
-    notifSection.appendChild(idleRow);
-
-    notifContent.appendChild(notifSection);
-
-    // ── Auto-Mute Workspace Patterns section ──────────────────────
-    const muteSection = document.createElement('div');
-    muteSection.className = 'settings-section';
-
-    const muteTitle = document.createElement('div');
-    muteTitle.className = 'settings-section-title';
-    muteTitle.textContent = 'Auto-Mute Workspace Patterns';
-    muteSection.appendChild(muteTitle);
-
-    const muteDesc = document.createElement('div');
-    muteDesc.className = 'settings-description';
-    muteDesc.textContent = 'Workspaces matching these patterns will have notifications automatically disabled. Use * as a wildcard.';
-    muteSection.appendChild(muteDesc);
-
-    const patternListContainer = document.createElement('div');
-    patternListContainer.className = 'mute-pattern-list';
-    muteSection.appendChild(patternListContainer);
-
-    function renderPatternList() {
-      patternListContainer.textContent = '';
-      const patterns = notificationStore.getMutedPatterns();
-      for (const pattern of patterns) {
-        const row = document.createElement('div');
-        row.className = 'shortcut-row';
-
-        const label = document.createElement('span');
-        label.className = 'shortcut-label';
-        label.style.fontFamily = "'Cascadia Code', Consolas, monospace";
-        label.textContent = pattern;
-        row.appendChild(label);
-
-        const removeBtn = document.createElement('button');
-        removeBtn.className = 'dialog-btn dialog-btn-secondary';
-        removeBtn.textContent = 'Remove';
-        removeBtn.style.fontSize = '11px';
-        removeBtn.style.padding = '2px 10px';
-        removeBtn.onclick = () => {
-          notificationStore.removeMutedPattern(pattern);
-          renderPatternList();
-        };
-        row.appendChild(removeBtn);
-
-        patternListContainer.appendChild(row);
-      }
-    }
-
-    renderPatternList();
-
-    // Add pattern row
-    const addPatternRow = document.createElement('div');
-    addPatternRow.className = 'shortcut-row';
-
-    const patternInput = document.createElement('input');
-    patternInput.type = 'text';
-    patternInput.className = 'notification-preset';
-    patternInput.placeholder = 'e.g. Agent *';
-    patternInput.style.flex = '1';
-    addPatternRow.appendChild(patternInput);
-
-    const addPatternBtn = document.createElement('button');
-    addPatternBtn.className = 'dialog-btn dialog-btn-primary';
-    addPatternBtn.textContent = 'Add';
-    addPatternBtn.style.fontSize = '11px';
-    addPatternBtn.style.padding = '2px 14px';
-    addPatternBtn.onclick = () => {
-      const val = patternInput.value.trim();
-      if (val) {
-        notificationStore.addMutedPattern(val);
-        patternInput.value = '';
-        renderPatternList();
-      }
-    };
-    addPatternRow.appendChild(addPatternBtn);
-
-    patternInput.onkeydown = (e) => {
-      if (e.key === 'Enter') {
-        e.preventDefault();
-        addPatternBtn.click();
-      }
-    };
-
-    muteSection.appendChild(addPatternRow);
-    notifContent.appendChild(muteSection);
-    dialog.appendChild(notifContent);
-
-    // ── Plugins tab content ─────────────────────────────────────
-    const pluginsContent = document.createElement('div');
-    pluginsContent.className = 'settings-tab-content';
-    tabContents['plugins'] = pluginsContent;
-
-    const installingPlugins = new Set<string>();
-
-    function renderPluginsTab() {
-      pluginsContent.textContent = '';
-
-      const registry = getPluginRegistry();
-      if (!registry) {
-        const msg = document.createElement('div');
-        msg.className = 'settings-description';
-        msg.textContent = 'Plugin system not initialized.';
-        pluginsContent.appendChild(msg);
-        return;
-      }
-
-      // ── Installed section ──
-      const installedTitle = document.createElement('div');
-      installedTitle.className = 'plugin-section-title';
-      installedTitle.textContent = 'Installed';
-      pluginsContent.appendChild(installedTitle);
-
-      const registeredPlugins = registry.getAllWithMeta();
-      if (registeredPlugins.length === 0) {
-        const msg = document.createElement('div');
-        msg.className = 'settings-description';
-        msg.textContent = 'No plugins installed.';
-        pluginsContent.appendChild(msg);
-      } else {
-        const installedGrid = document.createElement('div');
-        installedGrid.className = 'plugin-browse-grid';
-        for (const entry of registeredPlugins) {
-          const card = createPluginCard({
-            plugin: entry.plugin,
-            manifest: entry.manifest,
-            isBuiltin: entry.builtin,
-            isEnabled: pluginStore.isEnabled(entry.plugin.id),
-            isInstalled: true,
-            onToggle: (enabled) => {
-              registry.setEnabled(entry.plugin.id, enabled);
-              renderPluginsTab();
-            },
-            onUninstall: async () => {
-              try {
-                await pluginInstaller.uninstallPlugin(entry.plugin.id);
-                renderPluginsTab();
-              } catch (e) {
-                console.error('[Plugins] Uninstall failed:', e);
-              }
-            },
-          });
-          installedGrid.appendChild(card);
-        }
-        pluginsContent.appendChild(installedGrid);
-      }
-
-      // ── Browse section ──
-      const browseTitle = document.createElement('div');
-      browseTitle.className = 'plugin-section-title';
-      browseTitle.style.marginTop = '16px';
-      browseTitle.textContent = 'Browse';
-      pluginsContent.appendChild(browseTitle);
-
-      const installedIds = new Set(registeredPlugins.map(e => e.plugin.id));
-      const browseEntries = (registryData as PluginRegistryData).plugins.filter(
-        (entry: RegistryEntry) => !installedIds.has(entry.id)
-      );
-
-      if (browseEntries.length === 0) {
-        const msg = document.createElement('div');
-        msg.className = 'settings-description';
-        msg.textContent = 'All registry plugins are installed.';
-        pluginsContent.appendChild(msg);
-      } else {
-        const browseGrid = document.createElement('div');
-        browseGrid.className = 'plugin-browse-grid';
-        for (const entry of browseEntries) {
-          const card = createPluginCard({
-            registryEntry: entry,
-            isBuiltin: false,
-            isEnabled: false,
-            isInstalled: false,
-            installing: installingPlugins.has(entry.id),
-            onInstall: async () => {
-              const [owner, repo] = entry.repo.split('/');
-              if (!owner || !repo) return;
-              installingPlugins.add(entry.id);
-              renderPluginsTab();
-              try {
-                await pluginInstaller.installPlugin(owner, repo);
-              } catch (e) {
-                console.error('[Plugins] Install failed:', e);
-              } finally {
-                installingPlugins.delete(entry.id);
-                renderPluginsTab();
-              }
-            },
-          });
-          browseGrid.appendChild(card);
-        }
-        pluginsContent.appendChild(browseGrid);
-      }
-    }
-
-    renderPluginsTab();
-    dialog.appendChild(pluginsContent);
-
-    // ── Shortcuts tab content ───────────────────────────────────
-    const shortcutsContent = document.createElement('div');
-    shortcutsContent.className = 'settings-tab-content';
-    tabContents['shortcuts'] = shortcutsContent;
-
-    // Keyboard Shortcuts header with Reset All button
-    const kbHeader = document.createElement('div');
-    kbHeader.className = 'settings-header';
-
-    const kbTitle = document.createElement('div');
-    kbTitle.className = 'settings-section-title';
-    kbTitle.textContent = 'Keyboard Shortcuts';
-    kbHeader.appendChild(kbTitle);
-
-    const resetAllBtn = document.createElement('button');
-    resetAllBtn.className = 'dialog-btn dialog-btn-secondary';
-    resetAllBtn.textContent = 'Reset All';
-    resetAllBtn.onclick = () => {
-      keybindingStore.resetAll();
-      renderShortcuts();
-    };
-    kbHeader.appendChild(resetAllBtn);
-
-    shortcutsContent.appendChild(kbHeader);
-
-    // Search input for filtering shortcuts
-    let shortcutSearchQuery = '';
-    const shortcutSearchInput = document.createElement('input');
-    shortcutSearchInput.type = 'text';
-    shortcutSearchInput.className = 'notification-preset shortcut-search';
-    shortcutSearchInput.placeholder = 'Filter shortcuts...';
-    shortcutSearchInput.oninput = () => {
-      shortcutSearchQuery = shortcutSearchInput.value.toLowerCase();
-      renderShortcuts();
-    };
-    shortcutsContent.appendChild(shortcutSearchInput);
-
-    // Shortcuts container
-    const shortcutsContainer = document.createElement('div');
-    shortcutsContainer.className = 'settings-shortcuts';
-    shortcutsContent.appendChild(shortcutsContainer);
-
-    dialog.appendChild(shortcutsContent);
-
-    // Currently capturing element (if any)
-    let capturingBadge: HTMLElement | null = null;
-    let capturingAction: ActionId | null = null;
-    let captureHandler: ((e: KeyboardEvent) => void) | null = null;
-
-    function stopCapture() {
-      if (capturingBadge) {
-        capturingBadge.classList.remove('capturing');
-        capturingBadge.textContent = formatChord(
-          keybindingStore.getBinding(capturingAction!)
-        );
-      }
-      if (captureHandler) {
-        document.removeEventListener('keydown', captureHandler, true);
-        captureHandler = null;
-      }
-      capturingBadge = null;
-      capturingAction = null;
-    }
-
-    function startCapture(badge: HTMLElement, actionId: ActionId) {
-      // Stop any existing capture first
-      stopCapture();
-
-      capturingBadge = badge;
-      capturingAction = actionId;
-      badge.classList.add('capturing');
-      badge.textContent = 'Press a key...';
-
-      captureHandler = (e: KeyboardEvent) => {
-        e.preventDefault();
-        e.stopImmediatePropagation();
-
-        // Escape cancels capture
-        if (e.key === 'Escape') {
-          stopCapture();
-          return;
-        }
-
-        // Ignore bare modifier keys
-        if (['Control', 'Shift', 'Alt', 'Meta'].includes(e.key)) {
-          return;
-        }
-
-        const chord = eventToChord(e);
-
-        // Must have at least Ctrl or Alt modifier
-        if (!chord.ctrlKey && !chord.altKey) {
-          return;
-        }
-
-        // Check for conflicts
-        const conflict = keybindingStore.findConflict(chord, actionId);
-        if (conflict) {
-          const conflictDef = DEFAULT_SHORTCUTS.find((d) => d.id === conflict);
-          const conflictLabel = conflictDef?.label ?? conflict;
-          const proceed = confirm(
-            `"${formatChord(chord)}" is already bound to "${conflictLabel}".\n\nOverwrite? The conflicting shortcut will be reset to its default.`
-          );
-          if (!proceed) {
-            stopCapture();
-            return;
-          }
-          // Reset the conflicting binding to its default
-          keybindingStore.resetBinding(conflict);
-        }
-
-        keybindingStore.setBinding(actionId, chord);
-        stopCapture();
-        renderShortcuts();
-      };
-
-      document.addEventListener('keydown', captureHandler, true);
-    }
-
-    function renderShortcuts() {
-      shortcutsContainer.textContent = '';
-
-      const categories: ShortcutCategory[] = ['Terminal', 'Clipboard', 'Tabs', 'Split', 'Workspace', 'Scroll', 'Zoom', 'Voice', 'Debug'];
-
-      for (const category of categories) {
-        const defs = DEFAULT_SHORTCUTS.filter((d) => {
-          if (d.category !== category) return false;
-          if (!shortcutSearchQuery) return true;
-          const chord = formatChord(keybindingStore.getBinding(d.id));
-          return d.label.toLowerCase().includes(shortcutSearchQuery)
-            || chord.toLowerCase().includes(shortcutSearchQuery);
-        });
-        if (defs.length === 0) continue;
-
-        const section = document.createElement('div');
-        section.className = 'settings-section';
-
-        const sectionTitle = document.createElement('div');
-        sectionTitle.className = 'settings-section-title';
-        sectionTitle.textContent = category;
-        section.appendChild(sectionTitle);
-
-        for (const def of defs) {
-          const row = document.createElement('div');
-          row.className = 'shortcut-row';
-
-          const label = document.createElement('span');
-          label.className = 'shortcut-label';
-          label.textContent = def.label;
-          row.appendChild(label);
-
-          const badge = document.createElement('span');
-          badge.className = 'shortcut-binding';
-          if (keybindingStore.isCustom(def.id)) {
-            badge.classList.add('custom');
-          }
-          badge.textContent = formatChord(keybindingStore.getBinding(def.id));
-          badge.onclick = () => startCapture(badge, def.id);
-          row.appendChild(badge);
-
-          const resetBtn = document.createElement('button');
-          resetBtn.className = 'shortcut-reset';
-          resetBtn.textContent = 'Reset';
-          resetBtn.title = `Reset to ${formatChord(def.defaultChord)}`;
-          resetBtn.onclick = () => {
-            keybindingStore.resetBinding(def.id);
-            renderShortcuts();
-          };
-          row.appendChild(resetBtn);
-
-          section.appendChild(row);
-        }
-
-        shortcutsContainer.appendChild(section);
-      }
-    }
-
-    renderShortcuts();
-
-    // ── Remote tab content ────────────────────────────────────────
-    const remoteContent = document.createElement('div');
-    remoteContent.className = 'settings-tab-content';
-    tabContents['remote'] = remoteContent;
-
-    // Password section
-    const pwSection = document.createElement('div');
-    pwSection.className = 'settings-section';
-
-    const pwTitle = document.createElement('div');
-    pwTitle.className = 'settings-section-title';
-    pwTitle.textContent = 'Password';
-    pwSection.appendChild(pwTitle);
-
-    const pwDesc = document.createElement('div');
-    pwDesc.className = 'settings-description';
-    pwDesc.textContent = 'Set a fixed password for phone access. We recommend saving this in a password manager like 1Password.';
-    pwSection.appendChild(pwDesc);
-
-    const pwForm = document.createElement('form');
-    pwForm.action = '/remote-setup';
-    pwForm.method = 'POST';
-    pwForm.autocomplete = 'on';
-    pwForm.onsubmit = (e) => e.preventDefault();
-
-    // Hidden username field for password manager association
-    const pwUsername = document.createElement('input');
-    pwUsername.type = 'text';
-    pwUsername.name = 'username';
-    pwUsername.autocomplete = 'username';
-    pwUsername.value = 'godly-terminal';
-    pwUsername.style.display = 'none';
-    pwForm.appendChild(pwUsername);
-
-    const pwInputRow = document.createElement('div');
-    pwInputRow.className = 'shortcut-row';
-    pwInputRow.style.gap = '8px';
-
-    const pwInput = document.createElement('input');
-    pwInput.type = 'password';
-    pwInput.name = 'password';
-    pwInput.id = 'godly-remote-password';
-    pwInput.autocomplete = 'new-password';
-    pwInput.className = 'notification-preset';
-    pwInput.style.flex = '1';
-    pwInput.style.fontFamily = "'Cascadia Code', Consolas, monospace";
-    pwInput.placeholder = 'Enter or generate a password';
-    pwInput.setAttribute('passwordrules', 'minlength: 16; maxlength: 128; required: upper, lower, digit, special;');
-    pwInput.value = remoteSettingsStore.getPassword();
-    pwInputRow.appendChild(pwInput);
-
-    const pwShowBtn = document.createElement('button');
-    pwShowBtn.className = 'dialog-btn dialog-btn-secondary';
-    pwShowBtn.textContent = 'Show';
-    pwShowBtn.style.fontSize = '11px';
-    pwShowBtn.style.padding = '2px 10px';
-    pwShowBtn.style.minWidth = '50px';
-    pwShowBtn.onclick = () => {
-      if (pwInput.type === 'password') {
-        pwInput.type = 'text';
-        pwShowBtn.textContent = 'Hide';
-      } else {
-        pwInput.type = 'password';
-        pwShowBtn.textContent = 'Show';
-      }
-    };
-    pwInputRow.appendChild(pwShowBtn);
-    pwForm.appendChild(pwInputRow);
-
-    const pwButtonRow = document.createElement('div');
-    pwButtonRow.className = 'shortcut-row';
-    pwButtonRow.style.marginTop = '8px';
-    pwButtonRow.style.gap = '8px';
-
-    const generateBtn = document.createElement('button');
-    generateBtn.className = 'dialog-btn dialog-btn-secondary';
-    generateBtn.textContent = 'Generate Strong Password';
-    generateBtn.onclick = () => {
-      pwInput.value = generatePassword(100);
-      pwInput.type = 'text';
-      pwShowBtn.textContent = 'Hide';
-    };
-    pwButtonRow.appendChild(generateBtn);
-
-    const copyBtn = document.createElement('button');
-    copyBtn.className = 'dialog-btn dialog-btn-secondary';
-    copyBtn.textContent = 'Copy';
-    copyBtn.onclick = async () => {
-      if (!pwInput.value) return;
-      try {
-        await navigator.clipboard.writeText(pwInput.value);
-        const original = copyBtn.textContent;
-        copyBtn.textContent = 'Copied!';
-        setTimeout(() => { copyBtn.textContent = original; }, 1500);
-      } catch {
-        // Clipboard API may not be available
-      }
-    };
-    pwButtonRow.appendChild(copyBtn);
-
-    const pwSaveBtn = document.createElement('button');
-    pwSaveBtn.className = 'dialog-btn dialog-btn-primary';
-    pwSaveBtn.textContent = 'Save Password';
-    pwSaveBtn.onclick = () => {
-      remoteSettingsStore.setPassword(pwInput.value);
-      const original = pwSaveBtn.textContent;
-      pwSaveBtn.textContent = 'Saved!';
-      setTimeout(() => { pwSaveBtn.textContent = original; }, 1500);
-    };
-    pwButtonRow.appendChild(pwSaveBtn);
-    pwForm.appendChild(pwButtonRow);
-
-    pwSection.appendChild(pwForm);
-    remoteContent.appendChild(pwSection);
-
-    // Server section
-    const serverSection = document.createElement('div');
-    serverSection.className = 'settings-section';
-
-    const serverTitle = document.createElement('div');
-    serverTitle.className = 'settings-section-title';
-    serverTitle.textContent = 'Server';
-    serverSection.appendChild(serverTitle);
-
-    const portRow = document.createElement('div');
-    portRow.className = 'shortcut-row';
-
-    const portLabel = document.createElement('span');
-    portLabel.className = 'shortcut-label';
-    portLabel.textContent = 'Port';
-    portRow.appendChild(portLabel);
-
-    const portInput = document.createElement('input');
-    portInput.type = 'number';
-    portInput.className = 'notification-preset';
-    portInput.style.width = '100px';
-    portInput.min = '1024';
-    portInput.max = '65535';
-    portInput.value = String(remoteSettingsStore.getPort());
-    portInput.onchange = () => {
-      const port = parseInt(portInput.value);
-      if (port >= 1024 && port <= 65535) {
-        remoteSettingsStore.setPort(port);
-      }
-    };
-    portRow.appendChild(portInput);
-    serverSection.appendChild(portRow);
-
-    const autoStartRow = document.createElement('div');
-    autoStartRow.className = 'shortcut-row';
-
-    const autoStartLabel = document.createElement('span');
-    autoStartLabel.className = 'shortcut-label';
-    autoStartLabel.textContent = 'Start remote server on launch';
-    autoStartRow.appendChild(autoStartLabel);
-
-    const autoStartCheckbox = document.createElement('input');
-    autoStartCheckbox.type = 'checkbox';
-    autoStartCheckbox.className = 'notification-checkbox';
-    autoStartCheckbox.checked = remoteSettingsStore.getAutoStart();
-    autoStartCheckbox.onchange = () => {
-      remoteSettingsStore.setAutoStart(autoStartCheckbox.checked);
-    };
-    autoStartRow.appendChild(autoStartCheckbox);
-    serverSection.appendChild(autoStartRow);
-
-    // API Key row
-    const apiKeyRow = document.createElement('div');
-    apiKeyRow.className = 'shortcut-row';
-
-    const apiKeyLabel = document.createElement('span');
-    apiKeyLabel.className = 'shortcut-label';
-    apiKeyLabel.textContent = 'API Key';
-    apiKeyRow.appendChild(apiKeyLabel);
-
-    const apiKeyDisplay = document.createElement('span');
-    apiKeyDisplay.className = 'shortcut-binding';
-    apiKeyDisplay.style.fontFamily = "'Cascadia Code', Consolas, monospace";
-    apiKeyDisplay.style.fontSize = '11px';
-    const storedApiKey = remoteSettingsStore.getApiKey();
-    apiKeyDisplay.textContent = storedApiKey ? storedApiKey.slice(0, 8) + '...' : '(not set)';
-    apiKeyRow.appendChild(apiKeyDisplay);
-
-    const apiKeyGenBtn = document.createElement('button');
-    apiKeyGenBtn.className = 'dialog-btn dialog-btn-secondary';
-    apiKeyGenBtn.textContent = storedApiKey ? 'Regenerate' : 'Generate';
-    apiKeyGenBtn.style.fontSize = '11px';
-    apiKeyGenBtn.style.padding = '2px 10px';
-    apiKeyGenBtn.onclick = () => {
-      const newKey = generateApiKey();
-      remoteSettingsStore.setApiKey(newKey);
-      apiKeyDisplay.textContent = newKey.slice(0, 8) + '...';
-      apiKeyGenBtn.textContent = 'Regenerate';
-    };
-    apiKeyRow.appendChild(apiKeyGenBtn);
-    serverSection.appendChild(apiKeyRow);
-
-    remoteContent.appendChild(serverSection);
-
-    // Connection section
-    const connSection = document.createElement('div');
-    connSection.className = 'settings-section';
-
-    const connTitle = document.createElement('div');
-    connTitle.className = 'settings-section-title';
-    connTitle.textContent = 'Connection';
-    connSection.appendChild(connTitle);
-
-    const connDesc = document.createElement('div');
-    connDesc.className = 'settings-description';
-    connDesc.textContent = 'Run setup-phone.ps1 to start the remote server with an ngrok tunnel. The script will use your saved password, port, and API key.';
-    connSection.appendChild(connDesc);
-
-    const connInstructions = document.createElement('div');
-    connInstructions.className = 'settings-description';
-    connInstructions.style.fontFamily = "'Cascadia Code', Consolas, monospace";
-    connInstructions.style.fontSize = '12px';
-    connInstructions.style.marginTop = '8px';
-    connInstructions.textContent = 'pwsh scripts/setup-phone.ps1';
-    connSection.appendChild(connInstructions);
-
-    remoteContent.appendChild(connSection);
-
-    dialog.appendChild(remoteContent);
 
     // ── Info footer ──────────────────────────────────────────────
     const footer = document.createElement('div');
@@ -1425,7 +156,9 @@ export function showSettingsDialog(): Promise<void> {
 
     // ── Close handling ──────────────────────────────────────────
     const close = () => {
-      stopCapture();
+      for (const provider of registry) {
+        provider.onDialogClose?.();
+      }
       overlay.remove();
       resolve();
     };
@@ -1435,8 +168,9 @@ export function showSettingsDialog(): Promise<void> {
     };
 
     const escHandler = (e: KeyboardEvent) => {
-      // Only close on Escape when not capturing
-      if (e.key === 'Escape' && !capturingBadge) {
+      // Don't close if any tab is in a capturing/modal state
+      const anyCapturing = registry.some(p => p.isCapturing?.());
+      if (e.key === 'Escape' && !anyCapturing) {
         close();
         document.removeEventListener('keydown', escHandler);
       }
