+++ src/components/settings/index.ts
@@ -1,0 +1,17 @@
+import { registerSettingsTab } from './registry';
+import { ThemesTab } from './themes-tab';
+import { TerminalTab } from './terminal-tab';
+import { NotificationsTab } from './notifications-tab';
+import { PluginsTab } from './plugins-tab';
+import { ShortcutsTab } from './shortcuts-tab';
+import { RemoteTab } from './remote-tab';
+
+// Register in display order
+registerSettingsTab(new ThemesTab());
+registerSettingsTab(new TerminalTab());
+registerSettingsTab(new NotificationsTab());
+registerSettingsTab(new PluginsTab());
+registerSettingsTab(new ShortcutsTab());
+registerSettingsTab(new RemoteTab());
+
+export { getSettingsTabRegistry } from './registry';
